requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==2.4.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==2.5.0
defusedxml==0.7.0rc2
Django==3.2.2
django-email-users==1.1.0
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dropbox==11.28.0
flake8==3.8.4
graphene==2.1.8
graphene-django==2.15.0
graphene-subscriptions==1.0.2
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
httpie==2.4.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.7.0
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==9.0.1
ply==3.11
promise==2.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.8.1
PyJWT==2.3.0
pyOpenSSL==20.0.1
PySocks==1.7.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
Rx==1.6.1
service-identity==18.1.0
simplejson==3.17.6
singledispatch==3.6.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
stone==3.3.1
text-unidecode==1.3
Twisted==21.2.0
txaio==21.2.1
uritemplate==4.1.1
urllib3==1.26.4
whitenoise==6.0.0
zope.interface==5.2.0
